bump metrics-*

+++ src/sources/mysqld/info_schema_query_response_time.rs
@@ -61,7 +61,7 @@ async fn check_stats(pool: &MySqlPool) -> Result<bool, Error> {
 
 struct Statistic {
     time: f64,
-    count: u32,
+    count: u64,
     total: f64,
 }
 
@@ -72,7 +72,7 @@ impl<'r> FromRow<'r, MySqlRow> for Statistic {
             .trim()
             .parse::<f64>()
             .unwrap_or(0.0);
-        let count = row.try_get::<u32, _>("COUNT")?;
+        let count = row.try_get::<u64, _>("COUNT")?;
         let total = row
             .try_get::<String, _>("TOTAL")?
             .trim()
